+++ magda-tenant-api/src/createTenantsRouter.ts
@@ -4,16 +4,18 @@ import {
     MAGDA_TENANT_ID_HEADER,
     MAGDA_ADMIN_PORTAL_ID
 } from "magda-typescript-common/src/registry/TenantConsts";
-import { mustBeAdmin } from "magda-typescript-common/src/authorization-api/authMiddleware";
+import { requireUnconditionalAuthDecision } from "magda-typescript-common/src/authorization-api/authMiddleware";
 import {
     installStatusRouter,
     createServiceProbe
 } from "magda-typescript-common/src/express/status";
+import AuthDecisionQueryClient from "magda-typescript-common/src/opa/AuthDecisionQueryClient";
 
 export interface ApiRouterOptions {
     database: Database;
     jwtSecret: string;
     authApiUrl: string;
+    authDecisionClient: AuthDecisionQueryClient;
 }
 
 function hasAdminPortalId(req: express.Request): boolean {
@@ -33,6 +35,7 @@ function hasAdminPortalId(req: express.Request): boolean {
 
 export default function createApiRouter(options: ApiRouterOptions) {
     const database = options.database;
+    const authDecisionClient = options.authDecisionClient;
 
     const router: express.Router = express.Router();
 
@@ -44,43 +47,53 @@ export default function createApiRouter(options: ApiRouterOptions) {
 
     installStatusRouter(router, status);
 
-    router.use(mustBeAdmin(options.authApiUrl, options.jwtSecret));
-
-    router.get("/tenants", function (req, res) {
-        try {
-            if (hasAdminPortalId(req)) {
-                database.getTenants().then((tenants) => {
-                    res.json(tenants);
-                    res.status(200);
-                });
-            } else {
-                res.status(400);
-                res.setHeader("Content-Type", "plain/text");
-                res.send("Incorrect tenant ID");
+    router.get(
+        "/tenants",
+        requireUnconditionalAuthDecision(authDecisionClient, {
+            operationUri: "object/tenant/read"
+        }),
+        function (req, res) {
+            try {
+                if (hasAdminPortalId(req)) {
+                    database.getTenants().then((tenants) => {
+                        res.json(tenants);
+                        res.status(200);
+                    });
+                } else {
+                    res.status(400);
+                    res.setHeader("Content-Type", "plain/text");
+                    res.send("Incorrect tenant ID");
+                }
+            } catch (e) {
+                console.error(e);
+                res.status(500);
             }
-        } catch (e) {
-            console.error(e);
-            res.status(500);
         }
-    });
+    );
 
-    router.post("/tenants", async function (req, res) {
-        try {
-            if (hasAdminPortalId(req)) {
-                const tenantId = await database.createTenant(req.body);
-                res.json(tenantId);
-                res.status(201);
-            } else {
-                res.status(400);
-                res.setHeader("Content-Type", "plain/text");
-                res.send("Incorrect tenant ID.");
+    router.post(
+        "/tenants",
+        requireUnconditionalAuthDecision(authDecisionClient, {
+            operationUri: "object/tenant/create"
+        }),
+        async function (req, res) {
+            try {
+                if (hasAdminPortalId(req)) {
+                    const tenantId = await database.createTenant(req.body);
+                    res.json(tenantId);
+                    res.status(201);
+                } else {
+                    res.status(400);
+                    res.setHeader("Content-Type", "plain/text");
+                    res.send("Incorrect tenant ID.");
+                }
+            } catch (e) {
+                console.error(e);
+                res.status(500);
             }
-        } catch (e) {
-            console.error(e);
-            res.status(500);
+            res.end();
         }
-        res.end();
-    });
+    );
 
     return router;
 }